Move helpers test into the right file in the tests folder
